requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-WTF==1.1.1
WTForms==3.0.1
Flask-Mail==0.9.1
SQLAlchemy==2.0.21
alembic==1.12.0
Werkzeug==2.3.7
cryptography==41.0.4
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.4
openai==0.28.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
nltk==3.8.1
textblob==0.17.1
pytz==2023.3
bcrypt==4.0.1
email-validator==2.0.0
itsdangerous==2.1.2
gunicorn==21.2.0
sentry-sdk[flask]==1.32.0
pytest==7.4.2
pytest-flask==1.2.0
black==23.7.0
flake8==6.0.0